+++ test/components.js
@@ -1,0 +1,1 @@
+import '@shgysk8zer0/components/button/share.js';